Zork 2.3
Added Assert.cs but kept on getting a red squiggly under IsDirection so I ended up making the code as identical as possible to the assignment page... with no luck of getting rid of the squiggly. Turns out I had accidentally put the directions list in the rooms string. (Note to self mind the "{}"s)